refactor: validation test 코드 분리
